Αdd test for quad area function (#61)
* added test for test quad area

* remove placeholder tests trivial

* reformat with black

* Update tests/test_grid_generation.py


* Update tests/test_grid_generation.py


* Update tests/test_grid_generation.py


* Update tests/test_grid_generation.py


---------
