+++ .gitignore
@@ -5,4 +5,4 @@ _build
 regional_mom6/_version.py
 regional_mom6.egg-info
 .pytest_cache
-env
+.env
