--- tests/test_trivial.py
@@ -1,15 +1,0 @@
-import numpy as np
-import pytest
-from regional_mom6 import angle_between
-
-
-# placeholder trivial test test
-@pytest.mark.parametrize(
-    ("v1", "v2", "v3", "true_angle"),
-    [
-        ([1, 0, 0], [0, 1, 0], [0, 0, 1], np.pi / 2),
-        ([1, 0, 0], [1, 1, 0], [0, 1, 1], np.pi / 4),
-    ],
-)
-def test_angle_between(v1, v2, v3, true_angle):
-    assert np.isclose(angle_between(v1, v2, v3), true_angle)
